Don't stretch narrow images in handheld view
If the handheld CSS has been selected then setting images to be full-width
was stretching narrow images (but leaving their height alone), distorting
them. For example see the sample rendering images in this section:
http://developers.whatwg.org/states-of-the-type-attribute.html#range-state-%28type=range%29

So don't resize images that already fit.

Instead, set a maximum width so that large images get scaled down to fit,
and ensure the height stays in proportion with the width.

Credit to zcorpan for the CSS:
http://krijnhoetmer.nl/irc-logs/whatwg/20121206#l-362